+++ client/src/components/profile/ProfileMedia.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 
 const ProfileMedia = ({ media: { title, type, description } }) => {
     return (
-        <div>
+        <div className='bg-grey p'>
             <h3 className='text-dark'>{title}</h3>
             <p>{type}</p>
             <p>